+++ resource_fairy/changelog.md
@@ -1,6 +1,17 @@
 ﻿# 0.0.1
 
 ## Features
+- Changed the daily amount of resources to now be three options for each resource type: Method, Amount, and Multiplier (see Nexus page or config file for details)
+
+## Bug Fixes
+- Will now continue working when reloading games or switching game modes
+
+## In Progress
+- 
+
+# 0.0.1
+
+## Features
 - Initial release
 
 ## Bug Fixes